react-flask-app/api/requirements.txt
attrs==23.2.0
beautifulsoup4==4.12.3
blinker==1.8.2
bs4==0.0.2
certifi==2024.6.2
click==8.1.7
exceptiongroup==1.2.1
Flask==3.0.3
Flask-Cors==4.0.1
gunicorn==22.0.0
h11==0.14.0
idna==3.7
importlib_metadata==7.2.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
numpy==2.0.0
outcome==1.3.0.post0
packaging==24.1
pandas==2.2.2
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
selenium==4.21.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
trio==0.25.1
trio-websocket==0.11.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
Werkzeug==3.0.3
wsproto==1.2.0
zipp==3.19.2
